--- 1week/calc.js
@@ -1,3 +1,0 @@
-exports.add = function(a,b){
-    return a+b;
-};
