--- 1week/calc2.js
@@ -1,6 +1,0 @@
-var calc ={};
-calc.add =function(a,b){
-  return a+b;  
-};
-
-module.exports=calc;
