--- 1week/ch02_test1.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-    <meta charset="UTF-8">
-    <title>test</title>
-    </head>
-    
-    <body>
-    <h1>안녕하세요!</h1>
-    </body>
-
-</html>
