--- 1week/ch02_test7.js
@@ -1,4 +1,0 @@
-var os=require('os');
-
-console.log('hostname:'+os.hostname());
-console.log('memory :'+os.freemem());
