+++ 이나현/1week/calc.js
@@ -1,0 +1,3 @@
+exports.add = function(a,b){
+    return a+b;
+};
