+++ 이나현/1week/calc2.js
@@ -1,0 +1,6 @@
+var calc ={};
+calc.add =function(a,b){
+  return a+b;  
+};
+
+module.exports=calc;
