+++ 이나현/1week/ch02.js
@@ -1,0 +1,7 @@
+console.log('argv 속성의 파라미터 수 : ' +process.argv.length);
+console.dir(process.argv);
+
+
+process.argv.forEach(function(item,index){
+    console.log(index+':'+item);
+});
