+++ 이나현/1week/ch02_test5.js
